refactor(tui): consolidate form validation pattern in BaseModalDialog
Eliminates code duplication by extracting validation error display
pattern into reusable base class methods.

Changes:
- Added _show_validation_error() to BaseModalDialog
  - Displays error message in #error-message widget
  - Focuses the problematic widget
  - Safe for dialogs without validation (try/except protection)
- Added _clear_validation_error() to BaseModalDialog
  - Clears error message before validation runs
- Removed _show_validation_error() from AlgorithmSelectionScreen (now inherited)
- Updated AlgorithmSelectionScreen to use base class methods
- Updated TaskFormDialog to use base class methods
  - Replaced inline error_message.update() + focus() pattern
  - Removed unused Static import

Benefits:
- ~15 lines of duplicated code eliminated
- Consistent error display across all dialogs
- Easy to add new validated dialogs
- 100% backwards compatible with existing dialogs
- Well-documented with clear expectations
